+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 
-## [3.2.0] - 2022-01-30
+## [3.2.0] - 2022-06-05
+
+Better support for using clear-folder logic in a script.
 
 ### Changed
 - Use recursive method to delete and leave path to conditional deleting
+- Dependency update
+- Access to logic without handling cli process
+
+### Fixed
+- Fix typescript declaration
 
 
 ## [3.1.0] - 2022-01-20